rename: 디렉토리 계층 구조를 REST API와 웹으로 분리함
