+++ src/test/java/kr/binarybard/hireo/web/member/service/MemberServiceTest.java
@@ -9,13 +9,14 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.util.Optional;
-import kr.binarybard.hireo.auth.dto.SignUpRequest;
+import kr.binarybard.hireo.web.auth.dto.SignUpRequest;
 import kr.binarybard.hireo.exception.MemberNotFoundException;
 import kr.binarybard.hireo.fixture.LoginFixture;
 import kr.binarybard.hireo.fixture.MemberFixture;
-import kr.binarybard.hireo.member.domain.Member;
-import kr.binarybard.hireo.member.dto.MemberMapper;
-import kr.binarybard.hireo.member.repository.MemberRepository;
+import kr.binarybard.hireo.web.member.domain.Member;
+import kr.binarybard.hireo.web.member.dto.MemberMapper;
+import kr.binarybard.hireo.web.member.repository.MemberRepository;
+import kr.binarybard.hireo.web.member.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
